src/flowexport: relocate inline tests to flowexport_tests.rs
Same playbook. flowexport.rs: 483 → 352 LOC.
cargo build/test 92/878.
